requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
apache-airflow-providers-cohere==1.0.0
apache-airflow-providers-pinecone==1.0.0
apache-airflow-providers-weaviate==1.0.0
apache-airflow-providers-pgvector==1.0.0
apache-airflow-providers-opensearch==1.0.0
apache-airflow-providers-openai==1.0.0
apache-airflow-providers-amazon==8.12.0
matplotlib==3.8.1
seaborn==0.13.0
scikit-learn==1.3.2
pandas==1.5.3
numpy==1.26.2
adjustText==0.8
weaviate-client==3.25.3
